requirements.txt
aiogram
groq
python-docx
PyPDF2
fpdf
httpx[socks]